refactor: 상수, 유틸 분리
