+++ .eslintrc.json
@@ -5,7 +5,7 @@
   },
   "parser": "babel-eslint",
   "extends": [
-    "airbnb",
+    "airbnb-base",
     "plugin:prettier/recommended",
     "plugin:react/recommended"
   ],
