+++ src/components/App.js
@@ -8,14 +8,18 @@ import Main from "./Main";
 import Modal from "./Modal";
 
 import {
-  PRIZE_TABLE,
-  RANKINGS,
   LOTTO_PRICE,
   LOTTO_RANGE,
   LOTTO_LENGTH,
+  INITIAL_RESULT,
 } from "../constants";
 
-import { countMatchedNumbers, createDistinctRandomIntegers } from "../utils";
+import {
+  calculateEarningRate,
+  createDistinctRandomIntegers,
+  deepCopyJSONObject,
+  getRanking,
+} from "../utils";
 
 const Container = styled.div`
   display: flex;
@@ -23,56 +27,17 @@ const Container = styled.div`
   align-items: center;
 `;
 
-const initialResult = {
-  rankCount: {
-    [RANKINGS.RANKING1]: 0,
-    [RANKINGS.RANKING2]: 0,
-    [RANKINGS.RANKING3]: 0,
-    [RANKINGS.RANKING4]: 0,
-    [RANKINGS.RANKING5]: 0,
-    [RANKINGS.NO_PRIZE]: 0,
-  },
-  earningRate: 0,
-};
-
-const initialState = {
-  lottos: [],
-  isModalOpen: false,
-  lottoResult: initialResult,
-};
-
-const getRanking = (lottoNumber, winningNumber, bonusNumber) => {
-  const numOfMatched = countMatchedNumbers(lottoNumber, winningNumber);
-
-  switch (numOfMatched) {
-    case 3:
-      return RANKINGS.RANKING5;
-    case 4:
-      return RANKINGS.RANKING4;
-    case 5:
-      if (countMatchedNumbers(lottoNumber, [bonusNumber])) {
-        return RANKINGS.RANKING2;
-      }
-      return RANKINGS.RANKING3;
-    case 6:
-      return RANKINGS.RANKING1;
-    default:
-      return RANKINGS.NO_PRIZE;
-  }
-};
-
-const calculateEarningRate = (rankCount, price) => {
-  const totalPrize = Object.values(RANKINGS).reduce((acc, ranking) => {
-    return acc + rankCount[ranking] * PRIZE_TABLE[ranking].prize;
-  }, 0);
-
-  return Math.round(((totalPrize - price) / price) * 100);
-};
-
 export default class App extends Component {
   constructor(props) {
     super(props);
-    this.state = initialState;
+
+    this.initialState = {
+      lottos: [],
+      isModalOpen: false,
+      lottoResult: deepCopyJSONObject(INITIAL_RESULT),
+    };
+
+    this.state = JSON.parse(JSON.stringify(this.initialState));
     this.action = {
       createLottos: (lottoCount) => {
         const lottos = Array.from({ length: lottoCount }, () =>
@@ -87,10 +52,7 @@ export default class App extends Component {
       },
 
       updateLottoResult: (winningNumbers, bonusNumber) => {
-        const result = {
-          ...initialResult,
-          rankCount: { ...initialResult.rankCount },
-        };
+        const result = deepCopyJSONObject(INITIAL_RESULT);
 
         const price = this.state.lottos.length * LOTTO_PRICE;
 
@@ -113,7 +75,7 @@ export default class App extends Component {
       },
 
       clear: () => {
-        this.setState(initialState);
+        this.setState(this.initialState);
       },
     };
   }
