+++ src/constants/index.js
@@ -1,3 +1,8 @@
 export { RANKINGS, PRIZE_TABLE } from "./prizeTable";
-export { LOTTO_PRICE, LOTTO_RANGE, LOTTO_LENGTH } from "./lotto";
+export {
+  LOTTO_PRICE,
+  LOTTO_RANGE,
+  LOTTO_LENGTH,
+  INITIAL_RESULT,
+} from "./lotto";
 export { ERROR_MESSAGE, GUIDE_MESSAGE } from "./messages";
